split avp file by encoder type
